Remove cranelift_codegen_shared
