--- cranelift/codegen/shared/README.md
@@ -1,2 +1,0 @@
-This crate contains shared definitions for use in both `cranelift-codegen-meta` and `cranelift
--codegen`.
